+++ CHANGELOG.md
@@ -1,6 +1,14 @@
-## 0.8.0 - unreleased
+## 0.9.0 - unreleased
 
 ADDED:
+
+CHANGED:
+
+REMOVED:
+
+FIXED:
+
+## 0.8.0 - 2016-12-20
 
 CHANGED: